backend/requirements.txt
debugpy
django
django-admin
djangorestframework
django-environ
psycopg2
redis
gunicorn
